Move navToLocString logic from Header to model
